apps/scanner/requirements.txt
boto3==1.34.0
PyYAML==6.0.1
jmespath==1.0.1
rich==13.7.1
mysql-connector-python==8.3.0
pymysql==1.1.0
redis==5.0.3
firebase-admin==6.3.0
slack-sdk==3.27.1
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
psycopg2-binary==2.9.9
pymongo==4.6.3
tinydb==4.8.0
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
setuptools==69.0.3
patool==1.15.0
pydrive2==1.19.0
appdirs==1.4.4
tqdm==4.66.1
funcy==2.0
fsspec==2024.2.0
opencv-python==4.8.1.78
# numpy<2 <-- Removed to allow dependency resolution
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
anthropic>=0.40.0

# Strict Chunking & Routing for Unstructured Sources
spacy>=3.7.0
blingfire>=0.1.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

couchdb==1.2
slack_sdk>=3.0.0
gdown>=4.0.0

# New connectors (P3)
pymssql>=2.2.0
azure-storage-blob>=12.19.0
google-cloud-bigquery>=3.17.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
extract-msg>=0.48.0
pyarrow>=15.0.0
fastavro>=1.9.0
snowflake-connector-python>=3.7.0
confluent-kafka>=2.3.0
jira>=3.8.0
hubspot-api-client>=9.0.0
simple-salesforce>=1.12.5

# Observability & resilience (Phase 13)
prometheus_client>=0.20.0
tenacity>=8.2.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-flask>=0.45b0
